UI: Update Sparkle to 2.1.0
Updates the Sparkle version to the latest (at the time of commit).
This requires using the new SPUStandardUpdateController
instead of SUUpdater, since the latter is now a stub.

Most of the SPUUpdaterDelegate code remains the as the
previous SUUpdaterDelegate. However, since SUAppcastItem
has deprecated initWithDictionary, this will need to be replaced.