UI: arrange properties view for narrow layout
This adds an additional class member to describe whether or not
OBSPropertiesView is being presented in a narrow layout.

This also changes AddColor to remove the label, replacing it with one
button that also shows the colour. In the narrow layout, the button has
no text. A similar change was made for AddFont.

Finally, AddGroup no longer makes QGroupBox checkable but adds its own,
separate check box. This is to prepare for possible QRadioButton.

Arranging the preview at the left and fixing the properties view at
300px to the right allows for a larger preview and more visible property
widgets at once on smaller displays.