UI: Add OBS_GROUP_RADIO type
This allows group boxes to handle choices between two (or more) options.
Each group contains the configurable options for that choice. The reason
for adding this feature is to take advantage of the narrow properties
layout which, by virtue of the fact that more widgets can be shown at
once, can lay out properties without having to hide/show them. Rather,
groups can be enabled/disabled.